--- client-mod/src/client/java/net/momirealms/craftEngineFabricMod/client/util/BlockUtils.java
@@ -1,21 +1,0 @@
-package net.momirealms.craftEngineFabricMod.client.util;
-
-import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
-import net.minecraft.block.Block;
-import net.minecraft.client.color.world.BiomeColors;
-import net.minecraft.world.biome.FoliageColors;
-
-public class BlockUtils {
-    public static void registerColor(Block block) {
-
-        ColorProviderRegistry.BLOCK.register(
-                (state, world, pos, tintIndex) -> {
-                    if (world != null && pos != null) {
-                        return BiomeColors.getFoliageColor(world, pos);
-                    }
-                    return FoliageColors.DEFAULT;
-                },
-                block
-        );
-    }
-}
