+++ client-mod/src/client/java/net/momirealms/craftengine/fabric/client/CraftEngineFabricModClient.java
@@ -1,19 +1,21 @@
-package net.momirealms.craftEngineFabricMod.client;
+package net.momirealms.craftengine.fabric.client;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
 import net.minecraft.block.Block;
+import net.minecraft.client.color.world.BiomeColors;
 import net.minecraft.client.gui.screen.DisconnectedScreen;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.registry.Registries;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
-import net.momirealms.craftEngineFabricMod.client.config.ModConfig;
-import net.momirealms.craftEngineFabricMod.client.network.CraftEnginePayload;
-import net.momirealms.craftEngineFabricMod.client.util.BlockUtils;
+import net.minecraft.world.biome.FoliageColors;
+import net.momirealms.craftengine.fabric.client.config.ModConfig;
+import net.momirealms.craftengine.fabric.client.network.CraftEnginePayload;
 
 import java.nio.charset.StandardCharsets;
 
@@ -28,7 +30,7 @@ public class CraftEngineFabricModClient implements ClientModInitializer {
             if (id.getNamespace().equals(CraftEngineFabricModClient.MOD_ID)) {
                 BlockRenderLayerMap.INSTANCE.putBlock(block, RenderLayer.getCutoutMipped());
                 if (id.getPath().contains("leaves")) {
-                    BlockUtils.registerColor(block);
+                    registerColor(block);
                 }
             }
         });
@@ -56,5 +58,17 @@ public class CraftEngineFabricModClient implements ClientModInitializer {
                 }
             });
         });
+    }
+
+    public static void registerColor(Block block) {
+        ColorProviderRegistry.BLOCK.register(
+                (state, world, pos, tintIndex) -> {
+                    if (world != null && pos != null) {
+                        return BiomeColors.getFoliageColor(world, pos);
+                    }
+                    return FoliageColors.DEFAULT;
+                },
+                block
+        );
     }
 }
